// src/types.ts
export type GithubUser = {
  avatar_url: string;
  events_url: string;
  followers_url: string;
  following_url: string;
  gists_url: string;
  gravatar_id: string;
  html_url: string;
  id: number;
  login: string;
  node_id: string;
  organizations_url: string;
  received_events_url: string;
  repos_url: string;
  score: number;
  site_admin: boolean;
  starred_url: string;
  subscriptions_url: string;
  type: string;
  url: string;
}

export type GithubRepository = {
  id: number;
  name: string;
  description: string | null;
  owner: GithubUser;
  stargazers_count: number;
}

export type UsernameWithGithubRepositories = {
  owner: string;
  repositories: GithubRepository[];
}